+++ backend-rust/src/graphql_api.rs
@@ -20,15 +20,6 @@ mod token;
 mod transaction;
 mod transaction_metrics;
 
-// TODO remove this macro, when done with first iteration
-/// Short hand for returning API error with the message not implemented.
-macro_rules! todo_api {
-    () => {
-        Err(crate::graphql_api::ApiError::InternalError(String::from("Not implemented")))
-    };
-}
-pub(crate) use todo_api;
-
 use crate::{
     connection::ConnectionQuery,
     graphql_api::search_result::SearchResult,
@@ -153,6 +144,8 @@ pub struct ApiServiceConfig {
         default_value = "100"
     )]
     module_reference_contract_link_events_collection_limit: u64,
+    #[arg(long, env = "CCDSCAN_API_CONFIG_MODULE_CONNECTION_LIMIT", default_value = "100")]
+    module_connection_limit: u64,
     #[arg(long, env = "CCDSCAN_API_CONFIG_REWARD_CONNECTION_LIMIT", default_value = "100")]
     reward_connection_limit: u64,
     #[arg(