+++ backend-rust/src/graphql_api/baker.rs
@@ -2787,7 +2787,7 @@ impl Baker {
     ) -> ApiResult<connection::Connection<String, InterimTransaction>> {
         let config = get_config(ctx)?;
         let pool = get_pool(ctx)?;
-        let query = ConnectionQuery::<i64>::new(
+        let query = ConnectionQuery::<DescendingI64>::new(
             first,
             after,
             last,
@@ -2833,14 +2833,14 @@ impl Baker {
                 FROM transactions
                 WHERE transactions.sender_index = $5
                 AND type_account = ANY($6)
-                AND index > $1 AND index < $2
+                AND index < $1 AND index > $2
                 ORDER BY
                     CASE WHEN NOT $3 THEN index END DESC,
                     CASE WHEN $3 THEN index END ASC
                 LIMIT $4
             ) ORDER BY index DESC"#,
-            query.from,
-            query.to,
+            i64::from(query.from),
+            i64::from(query.to),
             query.is_last,
             query.limit,
             account_index,
@@ -2886,8 +2886,8 @@ impl Baker {
             .await?;
 
             connection.has_previous_page =
-                result.min_id.map_or(false, |db_min| db_min < page_min_id);
-            connection.has_next_page = result.max_id.map_or(false, |db_max| db_max > page_max_id);
+                result.max_id.map_or(false, |db_max| db_max > page_max_id);
+            connection.has_next_page = result.min_id.map_or(false, |db_min| db_min < page_min_id);
         }
 
         Ok(connection)
@@ -3162,10 +3162,16 @@ impl<'a> BakerPool<'a> {
         #[graphql(desc = "Returns the last _n_ elements from the list.")] last: Option<u64>,
         #[graphql(desc = "Returns the elements in the list that come before the specified cursor.")]
         before: Option<String>,
-    ) -> ApiResult<connection::Connection<DescendingI64, DelegationSummary>> {
+    ) -> ApiResult<
+        connection::Connection<NestedCursor<DescendingI64, DescendingI64>, DelegationSummary>,
+    > {
+        type StakeCursor = DescendingI64;
+        type BakerIdCursor = DescendingI64;
+        type Cursor = NestedCursor<StakeCursor, BakerIdCursor>;
+
         let pool = get_pool(ctx)?;
         let config = get_config(ctx)?;
-        let query = ConnectionQuery::<DescendingI64>::new(
+        let query = ConnectionQuery::<Cursor>::new(
             first,
             after,
             last,
@@ -3181,39 +3187,90 @@ impl<'a> BakerPool<'a> {
                     delegated_restake_earnings as restake_earnings,
                     delegated_stake as staked_amount
                 FROM accounts
-                WHERE delegated_target_baker_id = $5 AND
-                    accounts.index > $2 AND accounts.index < $1
+                WHERE delegated_target_baker_id = $7
+                    AND ((delegated_stake > $2 AND delegated_stake < $1)
+                    OR (delegated_stake = $2 AND index > $3)
+                    OR (delegated_stake = $1 AND index < $4))
                 ORDER BY
-                    (CASE WHEN $4 THEN accounts.index END) ASC,
-                    (CASE WHEN NOT $4 THEN accounts.index END) DESC
-                LIMIT $3
+                    (CASE WHEN $6 THEN delegated_stake END) ASC,
+                    (CASE WHEN $6 THEN accounts.index END) ASC,
+                    (CASE WHEN NOT $6 THEN delegated_stake END) DESC,
+                    (CASE WHEN NOT $6 THEN accounts.index END) DESC
+                LIMIT $5
             ) AS delegators
-            ORDER BY delegators.index DESC",
-            i64::from(query.from),
-            i64::from(query.to),
+            ORDER BY delegators.staked_amount DESC, delegators.index DESC",
+            i64::from(query.from.outer),
+            i64::from(query.to.outer),
+            i64::from(query.from.inner),
+            i64::from(query.to.inner),
             query.limit,
             query.is_last,
             self.id
         )
         .fetch(pool);
+
         let mut connection = connection::Connection::new(false, false);
         while let Some(delegator) = row_stream.try_next().await? {
-            connection.edges.push(connection::Edge::new(delegator.index.into(), delegator));
+            let cursor = NestedCursor {
+                outer: delegator.staked_amount.into(),
+                inner: delegator.index.into(),
+            };
+            connection.edges.push(connection::Edge::new(cursor, delegator));
         }
-        if let Some(page_max_index) = connection.edges.first() {
-            if let Some(max_index) = sqlx::query_scalar!(
-                "SELECT MAX(index) FROM accounts WHERE delegated_target_baker_id = $1",
-                self.id
+
+        let (Some(last_item), Some(first_item)) =
+            (connection.edges.last(), connection.edges.first())
+        else {
+            // No items so we just return without updating next/prev page info.
+            return Ok(connection);
+        };
+
+        let collection_ends = sqlx::query!(
+            "WITH
+                starting as (
+                    SELECT
+                        index AS start_index,
+                        delegated_stake AS start_stake
+                    FROM accounts
+                    WHERE
+                        delegated_target_baker_id = $1
+                    ORDER BY
+                        delegated_stake DESC,
+                        index DESC
+                    LIMIT 1
+                ),
+                ending as (
+                    SELECT
+                        index AS end_index,
+                        delegated_stake AS end_stake
+                    FROM accounts
+                    WHERE
+                        delegated_target_baker_id = $1
+                    ORDER BY
+                        delegated_stake ASC,
+                        index ASC
+                    LIMIT 1
+                )
+           SELECT * FROM starting, ending",
+            self.id
+        )
+        .fetch_optional(pool)
+        .await?;
+        let collection_ends = collection_ends.ok_or_else(|| {
+            ApiError::InternalError(
+                "Failed to find collection ends for a non-empty collection".to_string(),
             )
-            .fetch_one(pool)
-            .await?
-            {
-                connection.has_previous_page = max_index > page_max_index.node.index;
-            }
-        }
-        if let Some(edge) = connection.edges.last() {
-            connection.has_next_page = edge.node.index != 0;
-        }
+        })?;
+        let collection_start_cursor = Cursor {
+            outer: collection_ends.start_stake.into(),
+            inner: collection_ends.start_index.into(),
+        };
+        let collection_end_cursor = Cursor {
+            outer: collection_ends.end_stake.into(),
+            inner: collection_ends.end_index.into(),
+        };
+        connection.has_previous_page = collection_start_cursor < first_item.cursor;
+        connection.has_next_page = collection_end_cursor > last_item.cursor;
         Ok(connection)
     }