+++ frontend/src/components/molecules/NodeLink.vue
@@ -1,8 +1,8 @@
 ﻿<template>
 	<div class="inline-block">
 		<NodeIcon class="h-4 inline align-text-top" :class="iconClass" />
-		<LinkButton class="px-2" @click="() => handleOnClick(node.id)">
-			{{ node.nodeName }}
+		<LinkButton class="px-2" @click="() => handleOnClick(nodeId)">
+			{{ nodeName }}
 		</LinkButton>
 	</div>
 </template>
@@ -10,11 +10,11 @@
 <script setup lang="ts">
 import { useDrawer } from '~/composables/useDrawer'
 import LinkButton from '~/components/atoms/LinkButton.vue'
-import type { NodeStatus } from '~/types/generated'
 import NodeIcon from '~/components/icons/NodeIcon.vue'
 
 type Props = {
-	node: NodeStatus
+	nodeId: string
+	nodeName: string
 	iconClass?: string
 }